formatting oops

+++ phnx_io_ros/include/phnx_io_ros/roboteq.hpp
@@ -70,11 +70,11 @@ public:
 
         std::unique_lock lk{mtx};
 
-        // Scale output 
+        // Scale output
         // Might need to lower this with the new ESC (maybe 250? -berto)
-        
-        int16_t level = percent * 1000 * this->power_scale; // power_scale currently 0.2
-        level = std::clamp(level, -1000, 1000);
+
+        int16_t level = percent * 1000 * this->power_scale;  // power_scale currently 0.2
+        level = std::clamp(level, int16_t(-1000), int16_t(1000));
         // via
         // https://www.scribd.com/document/832439076/Roboteq-Controllers-User-Manual-v3-2-225-488
         // "!G 1 'x' _", -1000 <= x <= 1000, we just never demanded 1000 so this was never an issue
